edit travis deploy

+++ deploy.sh
@@ -3,6 +3,8 @@
 rm -rf node_modules
 npm install
 npm run build
-ssh root@vps.jarrekk.com "rm -rf /frontend/*"
-rsync -avz ./dist/* root@vps.jarrekk.com:/frontend/dist/
-ssh root@vps.jarrekk.com "docker restart backhend_nginx_1"
+if [ $? -eq 0 ]; then
+  ssh root@vps.jarrekk.com "rm -rf /frontend/*"
+  rsync -avz ./dist/* root@vps.jarrekk.com:/frontend/dist/
+  ssh root@vps.jarrekk.com "docker restart backhend_nginx_1"
+fi
